+++ projects/scion/dimension/src/lib/dimension.directive.ts
@@ -9,18 +9,18 @@
  */
 
 import { Directive, ElementRef, EventEmitter, Input, NgZone, OnDestroy, Output } from '@angular/core';
-import { asapScheduler, fromEvent, merge, Observable, Subject, timer } from 'rxjs';
-import { distinctUntilChanged, map, takeUntil } from 'rxjs/operators';
-
-export const NULL_DIMENSION: SciDimension = {offsetWidth: 0, offsetHeight: 0, clientWidth: 0, clientHeight: 0};
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
+import { captureElementDimension, SciDimension, SciDimensionService } from './dimension.service';
 
 /**
  * Allows observing changes to host element's size.
  *
- * This directive can be used in 'Default' or 'OnPush' change detection context.
+ * See {SciDimensionService} for more information.
  *
  * ---
  * Usage:
+ *
  * <div sciDimension (sciDimensionChange)="onDimensionChange($event)"></div>
  */
 @Directive({
@@ -33,7 +33,7 @@ export class SciDimensionDirective implements OnDestroy {
   private _destroy$ = new Subject<void>();
 
   /**
-   * Emits when the dimension of the host element changes.
+   * Upon subscription, it emits the host element's dimension, and then continuously emits when the dimension of the host element changes.
    */
   @Output('sciDimensionChange') // tslint:disable-line:no-output-rename
   public dimensionChange = new EventEmitter<SciDimension>();
@@ -45,19 +45,23 @@ export class SciDimensionDirective implements OnDestroy {
   @Input()
   public emitOutsideAngular: boolean;
 
-  constructor(host: ElementRef, private _ngZone: NgZone) {
-    this._host = host.nativeElement as HTMLElement;
+  constructor(host: ElementRef<HTMLElement>,
+              private _dimensionService: SciDimensionService,
+              private _ngZone: NgZone) {
+    this._host = host.nativeElement;
+    this.installDimensionListener();
+  }
 
-    // run outside of the Angular zone to not trigger app ticks on every event
+  private installDimensionListener(): void {
     this._ngZone.runOutsideAngular(() => {
-      this.dimensionChange$()
+      this._dimensionService.dimension$(this._host)
         .pipe(takeUntil(this._destroy$))
         .subscribe((dimension: SciDimension) => {
           if (this.emitOutsideAngular) {
-            this.emitDimensionChange(dimension);
+            this.dimensionChange.emit(dimension);
           }
           else {
-            this._ngZone.run(() => this.emitDimensionChange(dimension));
+            this._ngZone.run(() => this.dimensionChange.emit(dimension));
           }
         });
     });
@@ -67,42 +71,7 @@ export class SciDimensionDirective implements OnDestroy {
    * Returns the current dimension of its host element.
    */
   public get dimension(): SciDimension {
-    return {
-      clientWidth: this._host.clientWidth,
-      clientHeight: this._host.clientHeight,
-      offsetWidth: this._host.offsetWidth,
-      offsetHeight: this._host.offsetHeight,
-    };
-  }
-
-  /**
-   * Emits when the dimension of the host element changes.
-   *
-   * This is a workaround until it is possible to listen for element dimension changes natively.
-   * @see https://wicg.github.io/ResizeObserver/
-   */
-  private dimensionChange$(): Observable<SciDimension> {
-    NgZone.assertNotInAngularZone();
-
-    return merge(
-      timer(0, asapScheduler), // Notify after directive construction to emit the initial dimension
-      this._ngZone.onStable, // When the Angular zone gets stable the dimension of the host element might have changed.
-      fromEvent(window, 'resize'), // When receiving a window size change, the Angular zone is not necessarily involved.
-      fromEvent(window, 'orientationchange'), // When receiving a window orientation change, the Angular zone is not necessarily involved.
-    )
-      .pipe(
-        map(() => this.dimension),
-        distinctUntilChanged((a, b) => {
-          return a.clientWidth === b.clientWidth &&
-            a.clientHeight === b.clientHeight &&
-            a.offsetWidth === b.offsetWidth &&
-            a.offsetHeight === b.offsetHeight;
-        }),
-      );
-  }
-
-  private emitDimensionChange(dimension: SciDimension): void {
-    this.dimensionChange.emit(dimension);
+    return captureElementDimension(this._host);
   }
 
   public ngOnDestroy(): void {
@@ -110,12 +79,3 @@ export class SciDimensionDirective implements OnDestroy {
   }
 }
 
-/**
- * Emitted upon a host element size change.
- */
-export interface SciDimension {
-  offsetWidth: number;
-  offsetHeight: number;
-  clientWidth: number;
-  clientHeight: number;
-}